requirements.txt
flask==2.3.2
flask-login==0.6.2
flask-sqlalchemy==3.0.3
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
pdfkit==1.0.0
python-docx==0.8.11
numpy==1.26.4
pandas==2.2.3
apscheduler==3.10.1
python-dotenv==1.0.0
werkzeug==2.3.7